+++ crmeb/src/main/java/com/zbkj/crmeb/nsl/nslwxapp/dao/NslCweightMapper.java
@@ -16,6 +16,6 @@ import java.util.List;
 public interface NslCweightMapper extends BaseMapper<NslCweight> {
 
     //根据车辆id查找配重信息列表
-    List<NslCweight> getCweightListByCraneId(int craneid,long pageindex,long pagesize);
+    List<NslCweight> getCweightListByCraneId(int craneid,int pageindex,int pagesize);
 
 }
